+++ CHANGELOG.md
@@ -27,6 +27,9 @@
 * Include global permissions in package.json base permissions ([UIEUS-474](https://folio-org.atlassian.net/browse/UIEUS-474))
 * Migrate tag flag from mod-configuration to mod-settings ([UIEUS-475](https://folio-org.atlassian.net/browse/UIEUS-475))
 * Fix Monthpicker closing immediately when clicking ([UIEUS-477](https://folio-org.atlassian.net/browse/UIEUS-477))
+* Implement proper Focus Management for MonthPicker ([UIEUS-485](https://folio-org.atlassian.net/browse/UIEUS-485))
+* Fix Monthpicker ESC key propagation ([UIEUS-490](https://folio-org.atlassian.net/browse/UIEUS-490))
+* Fix clicking Monthpicker calendar icon does not close the calendar when already open ([UIEUS-491](https://folio-org.atlassian.net/browse/UIEUS-491))
 * Improve priority of ReportInfoButton icons ([UIEUS-492](https://folio-org.atlassian.net/browse/UIEUS-492))
 
 ## [11.0.1](https://github.com/folio-org/ui-erm-usage/tree/v11.0.1) (2025-04-11)